style enhancement for tag items
