+++ src/components/specific/bcf/bcf-topics-list/bcf-index-cell/BcfIndexCell.vue
@@ -13,7 +13,7 @@
 <script>
 import { computed } from "vue";
 
-import { adjustColor } from "@/components/specific/bcf/bcf-settings/adjustColor.js";
+import { adjustColor } from "@/components/generic/color-selector/colors.js";
 
 export default {
   props: {
