+++ src/components/specific/bcf/bcf-topics-list/open-topic-issue/OpenTopicIssue.vue
@@ -255,7 +255,7 @@
 <script>
 import { computed, ref } from "vue";
 import { useI18n } from "vue-i18n";
-import { adjustColor } from "@/components/specific/bcf/bcf-settings/adjustColor.js";
+import { adjustColor } from "@/components/generic/color-selector/colors.js";
 import { useBcf } from "@/state/bcf.js";
 import { useProjects } from "@/state/projects.js";
 import { useModels } from "@/state/models.js";
